+++ lib/state_flow.dart
@@ -10,4 +10,3 @@ export 'src/animation/animation_controller.dart';
 export 'src/widgets/state_flow_widget.dart';
 export 'src/state/state_value.dart';
 export 'src/state/state_flow_builder.dart';
-