+++ CHANGELOG.md
@@ -2,3 +2,13 @@
 
 * Initial release.
 Initial Release
+
+## 0.0.2
+
+* Easy Navigation added for navigating one screen to another screen.
+* Advanced Routing added.
+* Dependency not registering issues are fixed.
+* Initialization and disposal of StateFlowWidget is now handled using onInit and onDispose.
+* onInit and onDispose issues are fixed.
+* No more # in the URL.
+* Readme.md updated.